requirements.txt
Django==1.8
django-allauth==0.25.2
django-socketio==0.3.9
gevent==0.13.6
gevent-socketio==0.2.1
gevent-websocket==0.9.5
greenlet==0.4.9
oauthlib==1.1.2
python-openid==2.2.5
requests==2.10.0
six==1.10.0
sphinx-me==0.3
